+++ Mindspace_back/src/comment/comment.service.ts
@@ -10,7 +10,7 @@ import { Comment } from './entities/comment.entity';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UserService } from '../user/user.service';
 import { CommentMapper } from './dto/comment.mapper.dto';
-import { CommentSingleResponseDto } from './dto/comment-single-response.dto';
+import { CommentResponseDto } from './dto/comment-response.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { BoardService } from '../board/board.service';
 import { CustomCommentRepository } from './repository/comment.repository';
@@ -22,7 +22,7 @@ import { NotificationService } from '../notification/notification.service';
 import { User } from '../user/entities/user.entity';
 import { Board } from '../board/entities/board.entity';
 import { PutCommentDto } from './dto/put-comment.dto';
-import { CommentResponseDto } from './dto/comment-response-dto';
+import { CommentSingleResponseDto } from './dto/comment-single-response.dto';
 
 @Injectable()
 export class CommentService {
@@ -59,7 +59,7 @@ export class CommentService {
     userId: string,
     createCommentDto: CreateCommentDto,
     parentId?: number,
-  ): Promise<CommentResponseDto> {
+  ): Promise<CommentSingleResponseDto> {
     console.log(
       `[createComment] Started comment creation for board ${boardId} by user ${userId}`,
     );
@@ -140,7 +140,7 @@ export class CommentService {
       pagingParams,
     );
 
-    const transformedComments: CommentSingleResponseDto[] = await Promise.all(
+    const transformedComments: CommentResponseDto[] = await Promise.all(
       comments.data.map(async (comment) => {
         const replies = await this.customCommentRepository.findReplies(
           comment.id,
