--- Mindspace_back/src/comment/dto/comment-response-dto.ts
@@ -1,6 +1,0 @@
-export class CommentResponseDto {
-  id: number;
-  userNickname: string;
-  content: string;
-  updatedAt: Date;
-}
