+++ Mindspace_back/src/comment/dto/comment.mapper.dto.ts
@@ -3,7 +3,7 @@ import { formatDistanceToNow } from 'date-fns';
 import { ko } from 'date-fns/locale';
 import { CreateCommentDto } from './create-comment.dto';
 import { Comment } from '../entities/comment.entity';
-import { CommentSingleResponseDto } from './comment-single-response.dto';
+import { CommentResponseDto } from './comment-response.dto';
 import { User } from '../../user/entities/user.entity';
 import { Board } from '../../board/entities/board.entity';
 
@@ -35,7 +35,7 @@ export class CommentMapper {
   static commentToResponseDto(
     comment: Comment,
     userId: string,
-  ): CommentSingleResponseDto {
+  ): CommentResponseDto {
     const isDeletedComment = comment.content === '삭제된 댓글입니다.';
     return {
       id: comment.id,
